Extracing notes in a better way
There were 2 bugs in our previous implementation.

First, the way we were creating notes was wrong. I forgot the
case where a note can be played again when it's already being played.
That gave us an extra NOTE_OFF event that was unwanted. Issue was
fixed by introducing a stack to hold the timestamps.

Second issue is that we were getting the note durations wrong.
Instead of tick delta's we wanted fractional values (1 = quarter note,
0.5 = eighth note, etc). That was fixed by dividing the tick
delta by the number of ticks per quarter note.
Please enter the commit message for your changes. Lines starting

+++ cs30/Vivace/Note.pde
@@ -4,7 +4,7 @@ class Note {
   int value;
   PVector size;
   PVector position;
-  
+
   Note() {
     value = 0;
     size = new PVector(15, 30);
@@ -21,10 +21,10 @@ class Note {
 
 class UpcomingNote extends Note {
   long startTimestamp;
-  long duration;
+  float duration;
   int fallSpeed;
 
-  UpcomingNote(int value, long start, long duration) {
+  UpcomingNote(int value, long start, float duration) {
     super();
     fallSpeed = 0;
     startTimestamp = start;
@@ -33,6 +33,6 @@ class UpcomingNote extends Note {
   }
 
   void update() {
-    position.y -= fallSpeed; // TODO: consider delta time 
+    position.y -= fallSpeed; // TODO: consider delta time
   }
 }
